Create Task 6
A seemingly difficult task that is easy to understand. 

The number of subjects is taken as input from the user, and using for loop, we determine the range of how many times the user should input marks of each subject (in order). 

Right after that we store the marks and add them to the "total" variable used later to determine grades and average